django_project/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==5.0.2
django-picklefield==3.1
django-q2==1.6.1
djangorestframework==3.14.0
fastapi==0.109.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pyOpenSSL==24.0.0
pytz==2024.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.36.3
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2024.1
zope.interface==6.2
